requirements.txt
argon2-cffi==21.1.0
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
awsebcli==3.20.3
backports.zoneinfo==0.2.1
bcrypt==3.2.0
blessed==1.19.1
botocore==1.23.54
Brotli==1.0.9
cached-property==1.5.2
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.3
coverage==6.2
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.12
django-allauth==0.44.0
django-baton==2.3.0
django-debug-toolbar==3.2.4
django-grappelli==2.15.5
django-invitations==1.9.3
django-recaptcha==3.0.0
django-sslserver==0.22
djangorestframework==3.13.1
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
h11==0.13.0
idna==2.10
jmespath==0.10.0
jsonschema==3.2.0
newrelic==7.6.0.173
oauthlib==3.1.0
outcome==1.1.0
paramiko==2.9.2
passlib==1.7.4
pathspec==0.9.0
psycopg2==2.9.2
pycparser==2.20
PyJWT==2.3.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
selenium==4.1.0
semantic-version==2.8.5
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.4
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
wcwidth==0.1.9
websocket-client==0.59.0
whitenoise==6.0.0
wsproto==1.0.0
